requirements.txt
pygame==2.0.1
PyYAML~=5.4.1